feat(sdk/java): public/private exported/non exported fields (#9763)
* feat(sdk/java): add @Internal annotation for fields

Fields can now be private. No need to have them public anymore.

All non transient/static fields will be serialized by default and
registered as a field to the Dagger API.
If you want to have a field serialized but not exposed it to the
API (so a user can retrieve the value) then it needs to be annotated with

    @Internal

Signed-off-by: Yves Brissaud <gh@lgtd.io>

* refactor: remove unnecessary jsonbuilder code

variable was not used

Signed-off-by: Yves Brissaud <gh@lgtd.io>

* invert the logic from Internal annotation to Function

- public serialized fields are exposed by default
- private serialized fields needs a `@Function` annotation to be exposed

By exposed this means a default function with the name of the field will
be created by the API so that the value can be exposed to the user

Signed-off-by: Yves Brissaud <gh@lgtd.io>

* docs(sdk/java) add state docs

Signed-off-by: Yves Brissaud <gh@lgtd.io>

---------

Signed-off-by: Yves Brissaud <gh@lgtd.io>