fix: prevent double buildkit caching on remote git operations
This was *probably* causing flakes, and had the potential to cause big
issues.

Essentially, for remote operations we were caching twice:
- The DagOp for the `Git.tree` operation (with a cache key of the git ID)
- The underlying git operation (with a cache key of the git commit)
This was very confusing to think about, and probably could have resulted
in lots of weird cache issues.

The simple fix here is to just remove the extra DagOp layer when caching
remote git sources, and to take more manual control over DagOp caching
in this case.

Ideally! We would have core interfaces, which would then allow us to
have different specializations of `GitRef` and `GitRepository` at the
dagql handler level, so we wouldn't actually need this kind of refactor.
But it's not *super* trivial to add them, so this is quick hacky fix to
get the behavior working as desired.

Signed-off-by: Justin Chadwell <me@jedevc.com>