feat(sdk/java) add support for enums
User defined enums must have the `@Enum` annotation.
Then they can be used in modules.
This annotation is required to know this is an enum during the annotation
parsing phase, as the object might not yet be accessible.

It also ensures only one single enum is defined with the same name.

Technically this is done by updating the deserialization of JSON as
enum values must use the `valueOf(String)` method and can't be deserialized
straight away.
See https://jakarta.ee/specifications/jsonb/3.0/jakarta-jsonb-spec-3.0#enum

Signed-off-by: Yves Brissaud <gh@lgtd.io>