chore: refactor dagql middlewares into install hooks
This simplifies the dagql middlewares into install hooks - once the
object has been installed, a hook can make changes to the server
*afterwards*. Additionally, we add tests for this behavior, since it's a
bit fiddly.

The main logic for doing this was to solve the locking problem. What was
happening before this was:
- Attempt to install an object (takes the lock)
- Run a middleware for this installation
- The middleware attempts to install it's own objects, or modify
  existing objects (attempts to take the lock, and deadlocks because
  this thread already has it)

It wasn't really possible to untangle this without adding additional
thread unsafety: we need to hold the lock for the duration of the
install.

By refactoring, we can install the object, and then just hook -
importantly, the hook can't influence whether the object gets installed,
which allows us to release the lock before running the middleware, which
solves our weird dependency problem.

Signed-off-by: Justin Chadwell <me@jedevc.com>