cli: Add --cloud flag (EXPERIMENTAL)
WARNING: This is a highly **EXPERIMENTAL** feature. This change is the
first concrete step towards implementing something that one day all
Dagger Cloud users **might** be able to use.

This flag requires that the CLI:
- is connected to Dagger Cloud
- has an org set (e.g. `dagger login <YOUR_CLOUD_ORG>`)
- that org must have the `CLOUD_ENGINES` feature enabled (or be
  a `Dagger-Internal` org)

This flag provisions a Cloud Engine just-in-time, connects to it and
then runs all subsequent operations in it. ALL commands will work
exactly the same, just that they will run against this Cloud Engine
which is configured with:
- `16 vCPUs`
- `32GB RAM`
- `100GB NVMe storage`

NOTE: If the CLI is a development version, the Engine provisioned will
be `main` (a.k.a. latest development).

By default, the Engine shuts down after `30 minutes` of inactivity
(a.k.a. TTL). As soon as another command runs, the TTL resets to `30
minutes`.

State persists across Engines. After 30+ minutes of inactivity, running
another command will start a new Engine with the same state that the
last running Engine had. While there are a few scenarios which may
result in engines starting with no state, this should happen no more
than 1 in 10 runs.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>