ci: Introduce a simpler call action (#9877)
After all Alternative CI jobs started failing since #9673, this
introduces a separate `call-ci-alt-runner` action which is only used by
these alternative CI runners. We want something simpler that has a
separate lifecyle from the main `call` action that we use in the primary
CI runners.

The problem with the current `call` action logging config is that GitHub
Actions UI breaks, as well as the Namespace UI, so troubleshooting the
source of this issue is very difficult.

I even downloaded the logs locally, which did not make it easier to find
the issue. FTR:

    ~/Downloads/logs_35440011783 on ☁️
    ❯ ls -lah
    total 264M
    drwx------ 11 gerhard staff        352 Mar 10 19:24  .
    drwxr-xr-x 59 gerhard _lpoperator 1.9K Mar 10 19:24  ..
    -rw-rw-r--  1 gerhard staff        28M Nov 30  1979 '0_docs-lint-on-namespace-remote-engine _ remote-dagger-engine.txt'
    -rw-rw-r--  1 gerhard staff        15M Nov 30  1979 '1_docs-lint-on-namespace-local-engine _ local-dagger-engine.txt'
    -rw-rw-r--  1 gerhard staff        86M Nov 30  1979 '2_test-cli-engine-on-namespace-remote-engine _ remote-dagger-engine.txt'
    -rw-rw-r--  1 gerhard staff        30M Nov 30  1979 '3_sdk-go-on-namespace-remote-engine _ remote-dagger-engine.txt'
    -rw-rw-r--  1 gerhard staff        30M Nov 30  1979 '4_sdk-python-on-namespace-remote-engine _ remote-dagger-engine.txt'
    -rw-rw-r--  1 gerhard staff        31M Nov 30  1979 '5_sdk-typescript-on-namespace-remote-engine _ remote-dagger-engine.txt'
    -rw-rw-r--  1 gerhard staff        15M Nov 30  1979 '6_sdk-go-dev-on-namespace-local-engine _ local-dagger-engine.txt'
    -rw-rw-r--  1 gerhard staff        16M Nov 30  1979 '7_sdk-typescript-dev-on-namespace-local-engine _ local-dagger-engine.txt'
    -rw-rw-r--  1 gerhard staff        16M Nov 30  1979 '8_sdk-python-dev-on-namespace-local-engine _ local-dagger-engine.txt'

What's most peculiar about this issue is that the trace in Dagger Cloud
doesn't show any issues:
https://v3.dagger.cloud/dagger/traces/57d1391e843c1d8a7cfb473f60c034e2

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>
Signed-off-by: kpenfound <kyle@dagger.io>