fix: types management
Rework the way we deal with types. Extract the type management
in dedicated classes to add specific behavior depending on the
types.

This also fixes a lot of issues with List.
In short, when you have a `List<Something>` and try to deserialize
some JSON to it, it's not straightforward: `Something` disappears,
so you just have `List` and `List` is an interface:

- the fact `Something` disappear is an issue when the type is required
  to deserialize. For instance an enum is serialized as a string, but
  deserialized into a specific type. If we have `["low", "medium", "high"]`
  to deserialize and only the type `List` we don't know if it's a list of string
  or a list of the enum. And as the enums needs a special call to be set
  we can't do it.
- `List` is an interface. We can't set something as an interface, it
  has to be a concrete type. For instance we have to use an `ArrayList`.
  But `ArrayList` is not the defined type.

All in all, the idea is here, when we are deserializing, to not deserialize
as `List` but always as arrays. Then to convert to a `List`.

So for instance if the parameter type is `List<MyEnum>` we will deserialize
the value into a `MyEnum[]` then call `Arrays.toList` on it.
The usage of the array fixes the issues mentionned above, and from a JSON
perspective list or arrays are identical.

The new types exposes several methods, to know if it's a list or not,
and to expose the different code blocks we need for the entrypoint generation.
In particular a new method has been added when we want the class of a type
instead of just adding `.class` in the code generation.

Most of the `List` related behavior (to return or get a list) is covered
in the `enums` integration tests.

Signed-off-by: Yves Brissaud <gh@lgtd.io>