ci: Run testdev-cgroupsv2 in larger & faster CI runners (#9887)
* ci: Run testdev-cgroupsv2 in larger & faster CI runners

Add the Redis memory WARNING to the test condition, do not modify the
container - this might not work as expected when e.g. the file system is
read-only.

While at it, pin the CI runner Ubuntu version: `24.04` instead of
`latest`.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

* ci: Run other Engine & CLI testdev-* jobs in this new CI runner

Considering the ~3x speed increase for `testdev-cgroupsv2`, I am curious
to see how the other jobs behave...

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

---------

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>