extend dagger helm chart (#9845)
* extend dagger helm chart

Signed-off-by: Jeffrey Holm <jeff.holm@scale.com>

* fix chart lint

Signed-off-by: Jeffrey Holm <jeff.holm@scale.com>

* fix(helm): add namespacing for engine daemonset volumes

Otherwise, we can end up accidentally reusing the same volume if we
create two engine daemonsets (which can happen easily during testing).

Signed-off-by: Justin Chadwell <me@jedevc.com>

* tests(helm): add tests for setting ports

Signed-off-by: Justin Chadwell <me@jedevc.com>

* chore: ensure that engine doesn't start multiple listeners on same port

We need this otherwise the helm chart will manually specify the `--addr`
of `/run/dagger/engine.sock` - which will add it in duplicate to the
listeners.

So, we de-duplicate the addresses, which ensures this can't happen.

Signed-off-by: Justin Chadwell <me@jedevc.com>

* chore(helm): switch from /var/run/dagger to /run/dagger

Signed-off-by: Justin Chadwell <me@jedevc.com>

---------

Signed-off-by: Jeffrey Holm <jeff.holm@scale.com>
Signed-off-by: Justin Chadwell <me@jedevc.com>
Co-authored-by: Justin Chadwell <me@jedevc.com>