chore: add git attachable fallback and bump minimum client version (#9921)
* chore: add git attachable fallback

We haven't actually changed any details of the API - and it's pretty
easy to give this as a fallback so that we don't need to bump any engine
versions anywhere.

Signed-off-by: Justin Chadwell <me@jedevc.com>

* chore: bump minimum client version

This is the only version with the right client attachables - e.g. the
new allow llm prompting, and the renamed `Git` attachable.

Signed-off-by: Justin Chadwell <me@jedevc.com>

* tests: remove unneccesary version override in legacy_test.go

Signed-off-by: Justin Chadwell <me@jedevc.com>

---------

Signed-off-by: Justin Chadwell <me@jedevc.com>