ci: Make testdev-everything-else & testdev-module-runtimes smaller
Remove LLM split since 40% of the time is spent in compiling the dev
Engine. Now that we have removed a few larger suites from
testdev-everything-else, this one should go through without a hitch.

Reduce parallelism to 14, so that we give the 16 CPU instances some
headroom. This is not an exact science, but I am keen to check if it
makes things more stable, without trading off too much speed.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>