attempt to deflake TestPlatform/TestCacheMounts
This test has started to flake out more often recently again. Based on
test output it looks like the attempt to prevent the cache mount from
being pruned wasn't even actually running by the time the test
containers were running (which is possible because it works by running a
separate long running container in a different goroutine, so the
goroutine isn't guaranteed to be running at any particular time).

This attempts to deflake by having the long-running container work as a
service depedency, which guarantees it is running and preventing cache
mount pruning while the actual test containers are running.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>