llm: avoid exposing the model to "variables" concept (#9993)
* llm: remove readVariable tool

This was a halfway measure to allowing the LLM to discover vars that
were set for it. As-is, it confuses the model into unnecessary calls and
into trying to pass variables as arguments to tools, which is a bit too
much ambiguity.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>

* llm: remove all mention of variables

Let's just make everything 100% focused and strictly around IDs, so the
model never (lol) tries anything funny.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>

* woops, allow expanding prompt vars though

these are fine because the LLM never actually sees them

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>

---------

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>