ui: chain spans based on actor
currently LLM call loops are always double spaced and calls are always
unchained, which is a shame because they actually are chained calls.

now we'll consider spans chained as long as they have the same actor
emoji. it's a little overzealous but works out for now because the tool
calling scheme is always effectively chained anyway.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>