llm(sync): update in-place instead of cloning
This is called internally by any API that needs to un-lazy the LLM, but
that meant every un-lazying call site would end up running the loop,
instead of just the once.

Now we use a sync.Once to coordinate, and just allocate a new one each
time we clone, instead of having the dirty flag.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>