feat(sdk/elixir): add req graphql adapter
This changeset add the [`Req`](https://hexdocs.pm/req/Req.html) as a new
HTTP client backend for GraphQL aims to solve the concurrent issue on
the `:httpc` backend. [1]

This backend can enable by adding a `:req` as a dependency to the `mix.exs`
and configure the backend by adding the line below to the `config/config.exs`
of the module:

    config :dagger, client: Dagger.Core.GraphQLClient.Req

And also add a integration test to ensure this backend work as expected.

[1] https://discord.com/channels/707636530424053791/1133501257051553845/1349417869737459763.

Signed-off-by: Thanabodee Charoenpiriyakij <wingyminus@gmail.com>