fix: ensure that we use views with trailing slash methods (#10101)
When we mixed in slashes here, we *should* have ensured that we updated
the view. Otherwise, we could encounter the following scenario:
- Module A uses v0.16.0 (no trailing slashes), and calls
  `Directory.entries` on a directory.
- Module A uses v0.17.0 (has trailing slashes), and calls
  `Directory.entries` on the same directory.

If this all occurs within the same session, then these calls will
generate the same ID - and the wrong result will leak between these (in
a racy way).

The correct way to solve this problem is to ensure that these generate
different IDs (which we can do by manually using `dagql.AllVersion`).

Signed-off-by: Justin Chadwell <me@jedevc.com>