chore: remove unused gitdns
There was technically only *one* way to use this left - if you created a
container, bound a service, and then built a dockerfile in that
container.

However, there are no tests for this. We've discussed deprecating
`Container.build` for `Directory.dockerBuild` which has no such
functionality. So I think it should be safe to remove, and assume that
no one was using this.

Signed-off-by: Justin Chadwell <me@jedevc.com>