evals(BuildMulti): spell things out a bit more
the types of failures we're seeing at this point seem entirely due to
the model's inconsistent knowledge of the Go toolchain (i.e. where `go
build` puts things by default), which we don't really have control over
and will vary humongously model-to-model.

in other words, while it's impressive that some models figure everything
out with just a few words, it seems more practical for these evals to
target the lowest common denominator and explicitly assert for desired
behavior only when it's an important mechanic (like UndoChanges).

Signed-off-by: Alex Suraci <alex@dagger.io>