evals(WorkspacePattern): attempt to clarify
some models aren't getting the hint and are just recording findings of
their own, and then researching them. kind of makes sense - the API is
weird and ambiguous (does the number arg to `research` mean 'research
the finding i recorded' or 'return the 0th thing for me to ... record'?

now it's weird, but at least not ambiguous.

also renamed testspace to workspace, since that's what we're testing,
really

Signed-off-by: Alex Suraci <alex@dagger.io>