fix possible race condition initializing session attachables
Theoretically, before this change it seems possible for the session
attachables of a nested client to not be ready by the time the actual
nested exec process is running. The session attachables run in a child
process (in order to avoid conflicts around SIGCHLD handling with the
init and secret command providers), but that child process could
theoretically not have actually initialized its server by the time the
child process for the user's exec has started.

This would likely require a very overloaded system for the session child
process to not get scheduled for such a long time, but not impossible.

The fix here uses a pipe fd to synchronize the parent/child proc so that
the parent only starts the user exec once the session attachable server
is actually running.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>