fix: graceful failure when opening terminals
Previously, we were often returning invalid terminals from the
`OpenTerminal` function. This could happen because gRPC does not
actually wait for the call to `Session` to complete - instead, any
errors returned (like the one that reports that the user can't open a
terminal in a `--progress=plain` session) is actually returned on the
first call to `Recv`.

To handle this, we now send a void `Ready` message, and wait to receive
that - this allows us to correctly receive the error, which we can then
correctly propagate upwards to inform the user what really went wrong.

Signed-off-by: Justin Chadwell <me@jedevc.com>