fix: don't panic in cmd/init
Technically, we *are* error handling, we should take these errors and
print them nicely, and exit correctly.

We've had bug reports that mention the `panic`, and interpret it as
being a dagger internal error. This seems fair, `panic` should be
reserved for exceptional + unrecoverable scenarios (not ones that can be
caused by users by passing bad input, like trying to execute a program
that's not in the `PATH`).

Signed-off-by: Justin Chadwell <me@jedevc.com>