Revert "mark self arg required"
I think this is causing Gemini to generate malformed function calls.
Just a theory, based on the fact that all models I've ever tested are
super keen to forget to provide this argument, even when it's marked
'required' - maybe Gemini is the only one struct enough to verify
against the schema.

This reverts commit 08a1085a5dbe0c511ac20960dd8eae1bd313036e.

Signed-off-by: Alex Suraci <alex@dagger.io>