fix current module during initialization (#10213)
* fix current module during initialization

Module initialization (when we ask the runtime for what type defs it
has) previously got just enough data to support checking the current
module name, but basically nothing else. This is because the module is
not yet fully constructed during initialization.

This update results in as much of the module being available during init
as is possible. This should cover various cases like:
1. access to dag.CurrentModule().Source()
2. creating execs during module initializtion (which previously panicked
   because the instance ID was empty)

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* expand test coverage; cleanup old code made unneeded by change

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* simplify integ test

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* add changelog

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

---------

Signed-off-by: Erik Sipsma <erik@sipsma.dev>