fix `llm.withModel` usage from a module (#10230)
* fix `llm.withModel` usage from a module

`withModel` was not loading env vars from the main client in the same way
that the `llm` constructor was, so calling `withModel` didn't result in
any auth being loaded along with the reset model.

This updates `withModel` to load the router in the same way as the
constructor so API keys are loaded consistently.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* add changelog

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

---------

Signed-off-by: Erik Sipsma <erik@sipsma.dev>