chore: update netconfhttp to support dynamic dns
The idea is that this can be used per-request to determine which search
domains and resolver we need to use.

This is super useful for cases like in go-git, where we would need to
register a global resolver, but we don't actually know the DNS servers
we need to use for a request until later.

Signed-off-by: Justin Chadwell <me@jedevc.com>