feat: move http auth methods to top-level git
...and deprecate the standalone `WithAuthToken` and `WithAuthHeader`
methods.

Originally, this started as just an attempt to configure HTTP and SSH
authentication similarly - we wanted to standardize on doing it all in
the top-level constructor, or all using `With` helpers. I have a
preference for keeping at the top-level, so that a user can't mistakenly
do `Directory.AsGit().WithAuthToken()`, which wouldn't be used. However,
as I refactored the `git` implementation, I realized there's a bigger
inconsistency here.

Prior to this refactor, we were loading credentials entirely
inconsistently. While for SSH, we would avoid a potentially expensive
request to the client for credentials if we were already provided them,
for HTTP, we were unconditionally just fetching them. This resulted in a
ton of code sprawl, which, if we move everything to the constructor, is
vastly simplified.

Signed-off-by: Justin Chadwell <me@jedevc.com>