codegen never call dagger
Previously, the maven codegen plugin was calling dagger to get
the schema json file to generate code.
This was done using a specific query inside the java module.
While this is great from Java perspective, this creates a duplication
and especially this is a way to diverge from the codegen query that
will be used when generating code for modules.

This change makes the introspection schema file mandatory. It's now
required to generate one in advance, or to use the `sdk | java | generate`
command to update the java code.

That way we can think about implementing new features like the experimental
indication without to worry about the introspection schema file being wrong.

Signed-off-by: Yves Brissaud <gh@lgtd.io>