fix: ensure local commit check actually works
This checks wasn't actually working as desired - we weren't actually
checking that the commit was local (we need `^{commit}` and `--verify`
for that), but the functionality wasn't entirely broken, because we were
missing the trailing whitespace (so `skipFetch` was never true, so we
lost a little optimization here).

While we're here, we need to make this command non-fatal - a failure
here doesn't actually need to be marked as a failure in OTEL, so we can
skip past it.

Signed-off-by: Justin Chadwell <me@jedevc.com>