tests: skip source hut tests (#10289)
SourceHut looks to have added go-away, which adds a challenge to all
pages - including the `?go-get` query pages used by go tooling to
determine repos. This seems(?) to have entirely broken access to
`git.sr.ht` when using `GOPRIVATE` (to disable the go proxy).

This seems to be a breaking change that prevents all go modules on
source hut (and dagger modules since we use the same approach). In the
meantime, we disable our tests for this.

Signed-off-by: Justin Chadwell <me@jedevc.com>