fix: ensure git does not skip remote repo access
It was possible to skip remote repo access when resolving a remote repo
if *we already knew the commit*. This is actually a bad idea, since then
this would allow it to potentially entirely skip authentication:

- `GitRepository.commit` is called, so we don't do a resolution
- We access the shared remote ref, which already has that commit cloned
- We immediately return

This *wouldn't* be possible by using a ref (it's only happening with a
commit). One solution (taken here), is just to always resolve the repo,
even when we're just getting a commit. This checks for authentication
and avoids the issue.

Alternatively, we could maybe separate out shared remote refs by the
authentication they use, but this adds a lot of overhead - if we used
different credential combos, we'd need separate remote refs, which
doesn't seem worth it (at least for now).

Signed-off-by: Justin Chadwell <me@jedevc.com>