fix: typescript clien gen conflict with module
I noticed during a demo that if a module source is `.`
and  implements functions and also generate a client, the
`sdk/index.ts` file will not expose the generated client in the `sdk`
folder because it doesn't expect it to be there.

I changes the logic in the runtime to handle this specific edge case
and updated integrations tests.

This behaviour could be handled in the codegen binary ideally but
it's a quick fixes for now.

Signed-off-by: Tom Chauveau <tom@dagger.io>