ci: Add alternative CI Runners 4 (#10258)
This sets up an Alternative CI Runner that runs a subset of our tests in an
Engine wrapped in kata-containers (GKE).

The goal is to let this run for a week or two, and see how the metrics compare
to the other runtimes. We are especially interested in:

- queue times
- Engine warm-up
- cache test (the second run)
- total duration over >=20 runs expressed as P90 P95 P99

Power pair: Fabiano Fidêncio

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>