sdk/java: locked cache
Set the sharing mode of the maven cache volume to lock.
This ensure there's no concurrent write to the cache that should limit
the risk of corruption of maven files, as concurrent access to the maven
local repository is not supported.

Error looks like:

    [ERROR] Failed to execute goal
    org.apache.maven.plugins:maven-install-plugin:3.1.2:install
    (default-install) on project dagger-sdk-parent: Failed to install
    metadata io.dagger:dagger-sdk-parent/maven-metadata.xml:
    Could not parse metadata
    /root/.m2/repository/io/dagger/dagger-sdk-parent/maven-metadata-local.xml:
    in epilog non whitespace content is not allowed but got 0 (position:
    END_TAG seen ...</metadata>\n0... @13:2)

To lock the cache might have an impact in term of performances in case
of concurrent access, but it will keep good performances on single
access and ensure there's no corruption anymore.

Signed-off-by: Yves Brissaud <gh@lgtd.io>