tangential: fix panic from vault secret provider
I managed to hit this via new codepaths, but was pre-existing condition.
Vault provider would panic when a secret wasn't found in a path (e.g.
`/testsecret.foo` exists but not `/testsecret.bar`).

Fixed this and updated the test case, which seems to have been
accidentally asserting on the error message resulting from the panic
happening and the grpc connection getting closed suddenly.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>