chore: make `TestCrossSessionSecretURICaching` resilient to cache prunes (#10340)
This test started flaking in one PR. Didn't repro but looking at it
seems like it could be explained by the cache getting pruned while the
test is running (which would cause the assertions of cache re-use to
fail).

The update uses nested execs to run commands, which ensures that the
session stays open for the duration of the test case assertions and thus
should prevent the cache from getting pruned until the test is done.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>