ci: Move a bunch more jobs to Alt 2 ci runners (#10327)
* ci: Move a bunch more jobs to alt 2 ci runners

These have proven themselves up to 30-40% faster. A few of these jobs
can benefit from caching, so let's see how well they work out in
practice.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

* ci: Trace evals workflow

Also trigger the evals workflow when the workflow file changes,
otherwise we will not know if it still works correctly following a
change (such as the one that happened recently - `runs-on` changed).

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

* ci: Split testdev-interface from testdev-everything-else

It's currently the longest suite in this group, it deserves its own job.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

* ci: Split a few more interrelated test suites from everything-else

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

---------

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>