ci: alt-runners-4: Update the runner (#10345)
The current runners are giving us results around 4 minutes for the first
job.  Those specific runners are using nested virt atop of Intel SPR or
EMR.

We're now switching to using nested virt atop of AMD CPUs, as we're
getting results closer to 3m30s with those, instead of the Intel ones.

Let's switch the runners for now to use the AMD ones, while we keep
working on shaving the time down more and more with different
approaches. :-).

Signed-off-by: Fabiano Fidêncio <fidencio@northflank.com>