core: allow to define a moduleTypeDefs function
This is an optional new entry point on the module SDK.
It's designed to perform the type definition from the module like
the actual moduleRuntime when an empty string is passed as argument.

If the function doesn't exist, the moduleRuntime is used instead.

But this allows to decouple both of them so we can registry types
of a module without the runnable module itself because to build it
needs the types defined by the module

Signed-off-by: Yves Brissaud <gh@lgtd.io>