llm: list_available_methods -> list_methods
no longer filters out already-selected methods, since that seems a bit
too stateful. ideally the model calls this infrequently, only when it
needs to learn about a new type. this seems to help - it still gets
called twice in gemini-2.0-flash > BuildMulti, but before it was calling
it before just about every call.

Signed-off-by: Alex Suraci <alex@dagger.io>