tests: avoid flake from error caching
As we discovered recently, errors can cache between two sessions when
they're active at the same time. This is a pain, and should probably be
considered a bug.

However, these tests shouldn't be failing - we've had this behavior
*forever*, it's not a new failure. It's also not what we're testing, the
flakes just add to our CI noise - we're aware it's an issue, and we
should fix it (and add a test specifically for it).

Signed-off-by: Justin Chadwell <me@jedevc.com>