refactor: client gen dependency loading
Pass the `moduleSourceID` instead of a JSON file so it simplifies SDK
implementation and let it rely on the codegen.

If a moduleSourceID is passed to the codegen CLI,
it will fetch its dependencies and store it in the
codegen configuration.

Improve Go & SDK client codegen to only generate what is needed:
- If there's only remote dep, do not load any dagger.json, just serve
  git deps
- If there's local dependency, load the dagger.json, (throw an error or
  emit a warning if missing) and serve them
- If the module itself needs to be serve, load the dagger;json (same
  error handling as local) and serve it.

That way, all dependencies are sent to the codegen and then it can
adapt its behaviour.

Signed-off-by: Tom Chauveau <tom@dagger.io>