workflows: Change Alt Runners 4
After a long week of experimentations, I think we were able to trim down
a little bit the expected values, thus we're bringing here 2 different
"types" that should demonstrate such potential improvement.

Together with this change, we're trying to stick to run what the Alt
Runners 1 are running:
* doc-lint
* sdk-go
* run those in sequence (not parallel)

Ideally someone from the community side should be able to take over this
PR, and test this against the GHA already existent before getting it
pushed.

One type is commented out as we're doing more extensive experiments on
it.

Signed-off-by: Fabiano Fidêncio <fidencio@northflank.com>
Signed-off-by: Gerhard Lazu <gerhard@dagger.io>