fix: ensure dagger develop/publish do not take args
Running `dagger develop path/to/module` is an easy typo - however, it's
not valid, you need to use the `-m` flag. This didn't error out though,
we just accepted any additional args and ignored them.

We should explicitly forbid args to these functions if we're not using
them, which will help prevent typos.

Signed-off-by: Justin Chadwell <me@jedevc.com>