fix caching of fields that return lists (#10410)
* retain ID digest of both enumerable+elements

IDs with an nth element selected were overwriting the "leaf" selection
rather than appending a new selection of the nth element.

This meant that when we loaded elements from a list in an ID the ID
digest was different for every element of the list with no stable digest
for the list as a whole.

So I *think* the end result was that we'd repeatedly call the field that
returns the list *for every element in the list*. This is obviously not
ideal for performance, though it would have only impacted dagql caching
so far.

The fix here changes IDs to append a new selection for the nth element
on top of the original selection that returns the list. It then handles
this carefully in ID loading such that the list is first loaded once
with a stable id digest (and thus only called once thanks to dagql
caching), after which elements are selected from it.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* add unit test coverage

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* dedupe code and add explanatory comment

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* add object list coverage to viztest

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

---------

Signed-off-by: Erik Sipsma <erik@sipsma.dev>