Minor engine build optimizations (#10421)
* optimize go module binary building

Two minor optimizations that trim a few seconds from builds of binaries:
* only call `go list` if building a package path with a `...` wildcard
  in it; if there's no wildcards there's no reason to call go list
* skip synchronous pre-download of go modules; this happens as needed in
  the actual go build step so it's not clear why we were running this
  synchronously before each build *with a cache buster*

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* optimize dagger-engine module source loading

We were loading the entire git repo via a contextual dir arg, which took
like 7s on my machine, and only after that excluding files.

Now we just apply the excludes to the contextual dir, which reduces the
time to load to ~1s on my machine.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

---------

Signed-off-by: Erik Sipsma <erik@sipsma.dev>