refactor: client gen dependency loading (#10393)
Pass the `moduleSourceID` instead of a JSON file so it simplifies SDK
implementation and let it rely on the codegen.

If a moduleSourceID is passed to the codegen CLI,
it will fetch its dependencies and store it in the
codegen configuration.

Improve SDK client codegen to be simpler and more error proof

That way, all dependencies are sent to the codegen and then it can
adapt its behaviour.

Signed-off-by: Tom Chauveau <tom@dagger.io>