core: move error origin => link conversion to sdk
* look for ExtendedError everywhere, so internal core calls and client
  usage set up links properly
* preserve ExtendedError compliance in Dagger client so that module
  functions that return errors preserve their origins (even from
  ExecError)

on paper this looks like SDK specific work, but really it's because we
use the Go SDK telemetry code in the engine. modules errors are handled
generically in the engine, so other SDKs shouldn't need to implement
this - they should only need to preserve GraphQL error extensions, which
we already wanted prior to this change.

Signed-off-by: Alex Suraci <alex@dagger.io>