add support for using default engine GC policy during manual prune
Previously the core API for manually pruning the engine local cache only
supported pruning *all* entries. This adds an optional arg to enable
using the engine-wide default GC policy during these prunes.

The use case we have in mind is one where the automatic GC is disabled
and instead triggered manually by users (e.g. at a time of low-traffic
such that GC overhead is has minimal interference).

Ideally, the core API for managing the local cache would support the
full breadth of pruning options as is possible in the engine config, but
that's a bigger effort; the option added here is still an improvement in
the meantime.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>