idtui: fix data race in TestTelemetry
This is just the minimum viable change to fix the data race on `main`.

We currently lock in the _user_ of the `dagui.DB` instead of in the
`dagui.DB` itself; in practice this race doesn't occur since it's
guarded by the mutex in the `Frontend`.

It could be worth considering pushing this down into the `DB` but I've
been here before and previously decided it's better to let the `DB`
consume things as fast as possible and trust the caller to synchronize
more coarsely.

So, this change just continues the status quo and adds a lock to the
OTLP receiver in the tests.

Signed-off-by: Alex Suraci <alex@dagger.io>