engine: use sync log processor for clients
The OTel log SDK batch processor consumes an enormous amount of memory
when we use it for engine clients. Each client gets its own processor,
along with an additional processor for each ancestor. The memory usage
really adds up quickly: each processor has a queue, each log batch is
written to each parent, and each export does a slices.Copy of the batch
of records.

Now we'll just write log records to SQLite as soon as they're emitted,
synchronously, one at a time. For the evals, this cut the memory usage
in half. SQLite seems to take this just fine.

This change brought the heap usage for the evals down from 14.38GB to
7.43GB.

Signed-off-by: Alex Suraci <alex@dagger.io>